+++ drivers/rapidio/rio_cm.c
@@ -1896,7 +1896,7 @@ out:
  * riocm_cdev_ioctl() - IOCTL requests handler
  */
 static long
-riocm_cdev_ioctl(struct file *filp, unsigned int cmd, unsigned long arg)
+riocm_cdev_ioctl(struct file *filp, unsigned int cmd, uintptr_t arg)
 {
 	switch (cmd) {
 	case RIO_CM_EP_GET_LIST_SIZE:
